+++ client/src/App.js
@@ -25,48 +25,48 @@ import Review from "./Components/Review/Review.jsx";
 import Map from "./Components/Map/Map";
 
 function App() {
-	return (
-		<Box bg="#EDEDED" className="App">
-			<Navbar />
-			<Routes>
-				<Route
-					exact
-					path="/"
-					element={
-						<Box>
-							<SearchBar />
-							<Products />
-						</Box>
-					}
-				/>
-				<Route path="/login" element={<Registro />}>
-					<Route index element={<Login />} />
-					<Route path="signup" element={<Signup />} />
-					<Route path="logout" element={<Logout />} />
-					<Route path="newPassword" element={<ForgetPassword />} />
-					<Route path="newPassword/:token" element={<NewPass />} />
-					<Route path="confirmar/:id" element={<ConfirmarCuenta />} />
-				</Route>
-				<Route path="/admin" element={<HomeAdmin />}>
-					<Route index element={<AdministrarProductos />} />
-					<Route path="perfil" element={<EditarPerfil />} />
-					<Route path="cambiar-pass" element={<CambiasPassw />} />
-				</Route>
-				<Route path="/user" element={<HomeUser />}>
-					<Route path="perfil" element={<Pruevas />} />
-				</Route>
-				<Route path="/owner" element={<HomeOwner />}>
-					<Route index element={<AdministrarProductosOwner />} />
-					<Route path="perfil" element={<Pruevas />} />
-				</Route>
-				<Route path="/map" element={<Map />} />
-				<Route path="/terms" element={<TermsAndConditions />} />
-				<Route path="/detail/:id" element={<ProductDetail />} />
-				<Route path="/review" element={<Review />} />
-			</Routes>
-			<Footer />
-		</Box>
-	);
+  return (
+    <Box bg="#EDEDED" className="App">
+      <Navbar />
+      <Routes>
+        <Route
+          exact
+          path="/"
+          element={
+            <Box>
+              <SearchBar />
+              <Products />
+            </Box>
+          }
+        />
+        <Route path="/login" element={<Registro />}>
+          <Route index element={<Login />} />
+          <Route path="signup" element={<Signup />} />
+          <Route path="logout" element={<Logout />} />
+          <Route path="newPassword" element={<ForgetPassword />} />
+          <Route path="newPassword/:token" element={<NewPass />} />
+          <Route path="confirmar/:id" element={<ConfirmarCuenta />} />
+        </Route>
+        <Route path="/admin" element={<HomeAdmin />}>
+          <Route index element={<AdministrarProductos />} />
+          <Route path="perfil" element={<EditarPerfil />} />
+          <Route path="cambiar-pass" element={<CambiasPassw />} />
+        </Route>
+        <Route path="/user" element={<HomeUser />}>
+          <Route path="perfil" element={<Pruevas />} />
+        </Route>
+        <Route path="/owner" element={<HomeOwner />}>
+          <Route index element={<AdministrarProductosOwner />} />
+          <Route path="perfil" element={<Pruevas />} />
+        </Route>
+        <Route path="/map" element={<Map />} />
+        <Route path="/terms" element={<TermsAndConditions />} />
+        <Route path="/detail/:id" element={<ProductDetail />} />
+        <Route path="/review" element={<Review />} />
+      </Routes>
+      <Footer />
+    </Box>
+  );
 }
 
 export default App;
