+++ client/src/Components/Navbar/Navbar.jsx
@@ -1,16 +1,16 @@
 import React, { useState } from "react";
 import {
-	Box,
-	Text,
-	Button,
-	Flex,
-	Menu,
-	MenuButton,
-	MenuList,
-	MenuItem,
-	Image,
-	Portal,
-	Stack,
+  Box,
+  Text,
+  Button,
+  Flex,
+  Menu,
+  MenuButton,
+  MenuList,
+  MenuItem,
+  Image,
+  Portal,
+  Stack,
 } from "@chakra-ui/react";
 import { Link } from "react-router-dom";
 // import { useAuth } from "../../Context/AuthContext";
@@ -20,222 +20,222 @@ import { AiOutlineMenu } from "react-icons/ai";
 import { BsFillTrashFill } from "react-icons/bs";
 import { useSelector, useDispatch } from "react-redux";
 import {
-	add_quantity_in_cart_local_storage,
-	subtract_quantity_in_cart_local_storage,
-	delete_product_in_cart_local_storage,
+  add_quantity_in_cart_local_storage,
+  subtract_quantity_in_cart_local_storage,
+  delete_product_in_cart_local_storage,
 } from "../../Redux/Actions";
 
 function Navbar() {
-	const dispatch = useDispatch();
-	const { products_in_cart_local_storage } = useSelector((state) => state);
-	// console.log(products_in_cart_local_storage);
-	// const { currentUser, signout } = useAuth();
-	const [_width, set_width] = useState(window.frames.innerWidth);
-	window.addEventListener("resize", () => {
-		set_width(window.frames.innerWidth);
-	});
+  const dispatch = useDispatch();
+  const { products_in_cart_local_storage } = useSelector((state) => state);
+  // console.log(products_in_cart_local_storage);
+  // const { currentUser, signout } = useAuth();
+  const [_width, set_width] = useState(window.frames.innerWidth);
+  window.addEventListener("resize", () => {
+    set_width(window.frames.innerWidth);
+  });
 
-	// let localStorage_of_products = localStorage.getItem("productos_carrito");
-	// console.log(JSON.parse(localStorage_of_products));
+  // let localStorage_of_products = localStorage.getItem("productos_carrito");
+  // console.log(JSON.parse(localStorage_of_products));
 
-	// async function handleLogout() {
-	// try {
-	// 	await signout();
-	// } catch (error) {
-	// 	console.log(error);
-	// }
-	// }
+  // async function handleLogout() {
+  // try {
+  // 	await signout();
+  // } catch (error) {
+  // 	console.log(error);
+  // }
+  // }
 
-	return (
-		<Flex p="10px" justify={"space-between"} bg="#242525" color="#ECEDEC">
-			{_width <= 688 ? (
-				<Menu>
-					<MenuButton aria-label="Options" variant="outline">
-						<AiOutlineMenu />
-					</MenuButton>
-					<MenuList color={"#242525"}>
-						<MenuItem>
-							<Link to="/#">Inicio</Link>
-						</MenuItem>
-						<MenuItem>
-							<Link to="/#">Sucursal</Link>
-						</MenuItem>
-						<MenuItem>
-							<Link to="/#">Pedido</Link>
-						</MenuItem>
-						<MenuItem>
-							<Link to="/#">RapidoMore</Link>
-						</MenuItem>
-					</MenuList>
-				</Menu>
-			) : (
-				<>
-					<Flex justify={"space-around"} align="center">
-						<Box mx="15px">
-							<Link to="/#">Inicio</Link>
-						</Box>
-						<Box mx="15px">
-							<Link to="/#">Sucursal</Link>
-						</Box>
-						<Box mx="15px">
-							<Link to="/#">Pedido</Link>
-						</Box>
-						<Box mx="15px">
-							<Link to="/#">RapidoMore</Link>
-						</Box>
-					</Flex>
-					<Flex align={"center"} mx="30px">
-						<Link to="/signup">
-							<AiOutlineUser />
-						</Link>
-						<Box ml={4} color="#ECEDEC">
-							<AiOutlineHeart />
-						</Box>
-						<Box cursor={"pointer"} ml={4} color="#ECEDEC">
-							<Box>
-								{products_in_cart_local_storage.length > 0 &&
-									typeof products_in_cart_local_storage !== "string" && (
-										<Box
-											bg="#FE0100"
-											fontSize={"8px"}
-											textAlign="center"
-											borderRadius={"full"}
-											pos="absolute"
-											w="12px"
-											top="6px"
-											right="34px"
-										>
-											{products_in_cart_local_storage?.length}
-										</Box>
-									)}
-								<Menu autoSelect={false} closeOnSelect={false}>
-									<MenuButton
-										disabled={
-											products_in_cart_local_storage.length === 0 &&
-											typeof products_in_cart_local_storage !== "string"
-												? true
-												: false
-										}
-									>
-										<FiShoppingCart />
-									</MenuButton>
-									<Portal>
-										<MenuList w={"600px"}>
-											{products_in_cart_local_storage.length > 0 &&
-												typeof products_in_cart_local_storage !== "string" &&
-												products_in_cart_local_storage.map((elem, i) => {
-													return (
-														<MenuItem key={i} display="block">
-															<Flex
-																align="center"
-																justify="space-around"
-																mt={i > 0 && "12px"}
-															>
-																<Flex w="200px" justify={"center"}>
-																	<Image
-																		src={elem.imagen0}
-																		h="100px"
-																		borderRadius="3px"
-																	/>
-																</Flex>
-																<Stack>
-																	<Text ml="10px">
-																		{elem.nombre.slice(
-																			0,
-																			(elem.nombre.length * 40) / 100
-																		)}
-																	</Text>
-																	<Flex
-																		align={"center"}
-																		justify="space-between"
-																	>
-																		<Button
-																			borderRadius={"full"}
-																			w="2px"
-																			bg="#242525"
-																			color="#FFFF"
-																			_hover={{ bg: "#242525", color: "#FFFF" }}
-																			onClick={() =>
-																				dispatch(
-																					subtract_quantity_in_cart_local_storage(
-																						elem
-																					)
-																				)
-																			}
-																		>
-																			-
-																		</Button>
-																		<Text mx="10px">{elem.cantidad}</Text>
-																		<Button
-																			bg="#242525"
-																			borderRadius={"full"}
-																			color="#FFFF"
-																			_hover={{ bg: "#242525", color: "#FFFF" }}
-																			onClick={() =>
-																				dispatch(
-																					add_quantity_in_cart_local_storage(
-																						elem
-																					)
-																				)
-																			}
-																		>
-																			+
-																		</Button>
-																		<Text
-																			mx="10px"
-																			w={"95px"}
-																		>{`$ ${elem.precio.PesosArg}`}</Text>
-																		<Box
-																			onClick={() =>
-																				dispatch(
-																					delete_product_in_cart_local_storage(
-																						elem
-																					)
-																				)
-																			}
-																		>
-																			<BsFillTrashFill />
-																		</Box>
-																	</Flex>
-																</Stack>
-															</Flex>
-														</MenuItem>
-													);
-												})}
-											<MenuItem display="block">
-												<Flex align="center" justify="center" h="100px">
-													<Text mx="10px" fontWeight={"extrabold"}>
-														Total:{" "}
-														{products_in_cart_local_storage.length > 0 &&
-															products_in_cart_local_storage.reduce(
-																(a, b) =>
-																	Number(b.cantidad) <= 1
-																		? a + Number(b.precio.PesosArg)
-																		: a +
-																		  Number(b.precio.PesosArg) * b.cantidad,
-																0
-															)}
-													</Text>
-													<Button
-														mx="10px"
-														bg="#242525"
-														color="#FFFF"
-														_hover={{ bg: "#242525", color: "#FFFF" }}
-													>
-														Finalizar Compra
-													</Button>
-												</Flex>
-											</MenuItem>
-										</MenuList>
-									</Portal>
-								</Menu>
-							</Box>
-						</Box>
-					</Flex>
-				</>
-			)}
-		</Flex>
-	);
+  return (
+    <Flex p="10px" justify={"space-between"} bg="#242525" color="#ECEDEC">
+      {_width <= 688 ? (
+        <Menu>
+          <MenuButton aria-label="Options" variant="outline">
+            <AiOutlineMenu />
+          </MenuButton>
+          <MenuList color={"#242525"}>
+            <MenuItem>
+              <Link to="/#">Inicio</Link>
+            </MenuItem>
+            <MenuItem>
+              <Link to="/#">Sucursal</Link>
+            </MenuItem>
+            <MenuItem>
+              <Link to="/#">Pedido</Link>
+            </MenuItem>
+            <MenuItem>
+              <Link to="/#">RapidoMore</Link>
+            </MenuItem>
+          </MenuList>
+        </Menu>
+      ) : (
+        <>
+          <Flex justify={"space-around"} align="center">
+            <Box mx="15px">
+              <Link to="/#">Inicio</Link>
+            </Box>
+            <Box mx="15px">
+              <Link to="/#">Sucursal</Link>
+            </Box>
+            <Box mx="15px">
+              <Link to="/#">Pedido</Link>
+            </Box>
+            <Box mx="15px">
+              <Link to="/#">RapidoMore</Link>
+            </Box>
+          </Flex>
+          <Flex align={"center"} mx="30px">
+            <Link to="/login">
+              <AiOutlineUser />
+            </Link>
+            <Box ml={4} color="#ECEDEC">
+              <AiOutlineHeart />
+            </Box>
+            <Box cursor={"pointer"} ml={4} color="#ECEDEC">
+              <Box>
+                {products_in_cart_local_storage.length > 0 &&
+                  typeof products_in_cart_local_storage !== "string" && (
+                    <Box
+                      bg="#FE0100"
+                      fontSize={"8px"}
+                      textAlign="center"
+                      borderRadius={"full"}
+                      pos="absolute"
+                      w="12px"
+                      top="6px"
+                      right="34px"
+                    >
+                      {products_in_cart_local_storage?.length}
+                    </Box>
+                  )}
+                <Menu autoSelect={false} closeOnSelect={false}>
+                  <MenuButton
+                    disabled={
+                      products_in_cart_local_storage.length === 0 &&
+                      typeof products_in_cart_local_storage !== "string"
+                        ? true
+                        : false
+                    }
+                  >
+                    <FiShoppingCart />
+                  </MenuButton>
+                  <Portal>
+                    <MenuList w={"600px"}>
+                      {products_in_cart_local_storage.length > 0 &&
+                        typeof products_in_cart_local_storage !== "string" &&
+                        products_in_cart_local_storage.map((elem, i) => {
+                          return (
+                            <MenuItem key={i} display="block">
+                              <Flex
+                                align="center"
+                                justify="space-around"
+                                mt={i > 0 && "12px"}
+                              >
+                                <Flex w="200px" justify={"center"}>
+                                  <Image
+                                    src={elem.imagen0}
+                                    h="100px"
+                                    borderRadius="3px"
+                                  />
+                                </Flex>
+                                <Stack>
+                                  <Text ml="10px">
+                                    {elem.nombre.slice(
+                                      0,
+                                      (elem.nombre.length * 40) / 100
+                                    )}
+                                  </Text>
+                                  <Flex
+                                    align={"center"}
+                                    justify="space-between"
+                                  >
+                                    <Button
+                                      borderRadius={"full"}
+                                      w="2px"
+                                      bg="#242525"
+                                      color="#FFFF"
+                                      _hover={{ bg: "#242525", color: "#FFFF" }}
+                                      onClick={() =>
+                                        dispatch(
+                                          subtract_quantity_in_cart_local_storage(
+                                            elem
+                                          )
+                                        )
+                                      }
+                                    >
+                                      -
+                                    </Button>
+                                    <Text mx="10px">{elem.cantidad}</Text>
+                                    <Button
+                                      bg="#242525"
+                                      borderRadius={"full"}
+                                      color="#FFFF"
+                                      _hover={{ bg: "#242525", color: "#FFFF" }}
+                                      onClick={() =>
+                                        dispatch(
+                                          add_quantity_in_cart_local_storage(
+                                            elem
+                                          )
+                                        )
+                                      }
+                                    >
+                                      +
+                                    </Button>
+                                    <Text
+                                      mx="10px"
+                                      w={"95px"}
+                                    >{`$ ${elem.precio.PesosArg}`}</Text>
+                                    <Box
+                                      onClick={() =>
+                                        dispatch(
+                                          delete_product_in_cart_local_storage(
+                                            elem
+                                          )
+                                        )
+                                      }
+                                    >
+                                      <BsFillTrashFill />
+                                    </Box>
+                                  </Flex>
+                                </Stack>
+                              </Flex>
+                            </MenuItem>
+                          );
+                        })}
+                      <MenuItem display="block">
+                        <Flex align="center" justify="center" h="100px">
+                          <Text mx="10px" fontWeight={"extrabold"}>
+                            Total:{" "}
+                            {products_in_cart_local_storage.length > 0 &&
+                              products_in_cart_local_storage.reduce(
+                                (a, b) =>
+                                  Number(b.cantidad) <= 1
+                                    ? a + Number(b.precio.PesosArg)
+                                    : a +
+                                      Number(b.precio.PesosArg) * b.cantidad,
+                                0
+                              )}
+                          </Text>
+                          <Button
+                            mx="10px"
+                            bg="#242525"
+                            color="#FFFF"
+                            _hover={{ bg: "#242525", color: "#FFFF" }}
+                          >
+                            Finalizar Compra
+                          </Button>
+                        </Flex>
+                      </MenuItem>
+                    </MenuList>
+                  </Portal>
+                </Menu>
+              </Box>
+            </Box>
+          </Flex>
+        </>
+      )}
+    </Flex>
+  );
 }
 
 export default Navbar;
